* Make value system work

+++ Blackjack/Models/Player.cs
@@ -3,5 +3,5 @@
 public class Player(string name)
 {
     public string Name { get; init; } = name;
-    public int Points { get; set; } = 1000;
+    public double Points { get; set; } = 1000f;
 }
